WIP: Extract SchedulingManager and move UN gateway to Services.
Move BG processing and inactivity-reminder policy off AppManager onto SchedulingManager; rename the notification adapter to UserNotificationCenterService in the Services layer.
